added a running banner with text in it
the banner  color is google blue , the text inside it is white. the text is supposed to be a link but for some reason it turns into purple and I can't override the colors with my css . So i put the text inside a <p> tag instead of the <a> tag. To get the link working change the <p> to <a> in line 203 .